ISSUE #4841 - extract thumbnail
